+++ tests/test_simulator.py
@@ -61,8 +61,9 @@ def test_every_simulated_event_is_actually_emitted_by_the_integration():
         "soccer_live_goal_cancelled", "soccer_live_kickoff_changed",
         "soccer_live_venue_changed", "soccer_live_opponent_changed",
         "soccer_live_substitution", "soccer_live_lineup_available",
-        "soccer_live_watchlist_event",
-    }
+            "soccer_live_watchlist_event",
+            "soccer_live_race_milestone",
+        }
     emitted = state_and_detail | set(_load("match_contract").PHASE_EVENTS.values())
     for sim_type, (bus_event, _phase) in MODULE.EVENT_TYPES.items():
         assert bus_event in emitted, (